+++ tests/test_baseline.py
@@ -11,7 +11,6 @@ import baseline_noise
 from dataclasses import asdict
 from calibration import CalibrationResult
 import baseline
-from baseline_utils import subtract_baseline_counts
 from radon.baseline import subtract_baseline_counts
 from fitting import FitResult, FitParams
